chore(tests): use plugin collection in tests/extract_range.rs
